pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
]
